+++ internal/data/query.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/dgraph-io/travel/internal/advisory"
-	"github.com/dgraph-io/travel/internal/places"
+	"github.com/dgraph-io/travel/internal/feeds/advisory"
+	"github.com/dgraph-io/travel/internal/feeds/places"
+	"github.com/dgraph-io/travel/internal/feeds/weather"
 	"github.com/dgraph-io/travel/internal/platform/graphql"
-	"github.com/dgraph-io/travel/internal/weather"
 	"github.com/pkg/errors"
 )
