+++ internal/data/store_test.go
@@ -5,10 +5,10 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dgraph-io/travel/internal/advisory"
-	"github.com/dgraph-io/travel/internal/places"
+	"github.com/dgraph-io/travel/internal/feeds/advisory"
+	"github.com/dgraph-io/travel/internal/feeds/places"
+	"github.com/dgraph-io/travel/internal/feeds/weather"
 	"github.com/dgraph-io/travel/internal/platform/tests"
-	"github.com/dgraph-io/travel/internal/weather"
 	"github.com/google/go-cmp/cmp"
 )
